% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_performance.R
\name{aggregate_performance}
\alias{aggregate_performance}
\alias{print.mscagg}
\title{Aggregate performance measures by cohort}
\usage{
aggregate_performance(perf.estimates, reference = NULL,
  design.levels = LETTERS)

\method{print}{mscagg}(x, ...)
}
\arguments{
\item{perf.estimates}{A set of performance estimates of class \code{mscraw}, as computed by \code{\link{compute_performance}}}

\item{reference}{The name of the reference score (default NULL, the first score is the reference). This will be the reference level of the scores in any later models.}

\item{design.levels}{A character vector of alternate short names for the scores, to be used in naming the designs. Default is \code{\link{LETTERS}}, so that a possible design would be \code{AB} or \code{ACD}, instead of \code{score1score2} or \code{score1score3score4}. The design variable is used in the definition of the random effects. See \code{\link{consistency}}.}
}
\value{
A list with the following elements, of class \code{mscagg}
\describe{
  \item{cohorts}{First item}
  \item{yi}{Second item}
  \item{vi}{First item}
  \item{contr}{Second item}
  \item{design}{First item}
  \item{design.matrix}{First item}
  \item{lbl}{Second item}
  \item{fn}{First item}
  \item{scores}{Second item}
  \item{ref}{The reference score}
}
}
\description{
Aggregate performance measures by cohort
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print basic aggregated performance measures
}}

\examples{
dat <- msc_sample_data()
bssamp <- get_bs_samples(dat, id, cohort, outcome, n.samples = 10, a, b, c, d, e)
perf <- compute_performance(bssamp, fn = calibration_slope, lbl = "CS")
agg <- aggregate_performance(perf)
agg
}
