% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casemix.r
\name{casemix}
\alias{casemix}
\title{Summary of case-mix for key moderators}
\usage{
casemix(bss, mods = NULL, max.levels = 20, plot = TRUE)
}
\arguments{
\item{bss}{output from \code{\link{get_bs_samples}}}

\item{mods}{Which variables should be examined, of the \code{mods} in \code{bss}? (default NULL, means all)}

\item{max.levels}{If a numeric variable has fewer than \code{max.levels} distinct values, than it will be displayed as a categorical / factor variable. If a categorical variable}

\item{plot}{if TRUE (the default) then summary plots are produced. If not, tables of summary statistics are returned.}
}
\description{
Examines case-mix according to key moderators, by study.
}
\examples{
dat <- msc_sample_data()
dat$agegrp <- cut(dat$age, c(0, 20, 40, 60, 80, 100))
dat$agegrp2 <- cut(dat$age, seq(0, 100, 5))
dat$agegrp3 <- as.numeric(dat$agegrp2)
bssamp <- get_bs_samples(dat, id, study, outcome, n.samples = 10, 
                  scores = c("a", "b", "c", "d", "e", "f"), 
                  moderators = c("age", "agegrp", "agegrp2", "agegrp3", "female", "sex", "x1"))
casemix(bssamp)
casemix(bssamp, plot = FALSE)
casemix(bssamp, mods = c("agegrp2", "agegrp3"), max.levels = 10)   

}
