% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_transitivity}
\alias{check_transitivity}
\alias{check_homogeneity}
\alias{check_consistency}
\title{Check assumptions of transitivity, consistency and homogeneity}
\usage{
check_transitivity(object, graph = TRUE)

check_homogeneity(object, dig = 3)

check_consistency(object, graph = TRUE)
}
\arguments{
\item{object}{An object from \code{msc}. For \code{check_consistency}, it must have been run with options \code{direct = TRUE, indirect = TRUE}}

\item{graph}{Should the results be displayed graphically? (default TRUE) If not, the coefficient table is returned.}

\item{dig}{Number of digits to be printed, see \code{\link{format.pval}}.}
}
\value{
A nested list (performance measure, then moderator, than score contrast) containing linear regression model output of the effect size (outcome) against the cohort-specific average moderator value.

A list containing datasets with tau-square, the Q likelihood ratio statistic, its degrees of freedom and its p-value.

A graph of the direct and indirect estimates for the various differences in score performance.
}
\description{
Check assumptions of transitivity, consistency and homogeneity
}
\section{Functions}{
\itemize{
\item \code{check_transitivity()}: Check assumption of transitivity

\item \code{check_homogeneity()}: Check assumption of homogeneity

\item \code{check_consistency()}: Check assumption of consistency

}}
\examples{
dat <- msc_sample_data()
mod <- msc(scores = c("a", "b", "c", "d",),
            cohort = "cohort", outcome = "mort3", subjid = "id",
            fn = list(`logit AUC` = c_statistic, 
            `calib slope` = calibration_slope), 
            model = "consistency", 
            direct = FALSE, indirect = FALSE,
            ref = "first", data = dat)
\dontrun{check_consistency(mod)
modm <- msc(scores = c("a", "b", "c", "d",),
            cohort = "cohort", outcome = "mort3", subjid = "id",
            fn = list(`logit AUC` = c_statistic, 
            `calib slope` = calibration_slope), 
            model = "consistency", 
            mods = c("age", "female", "x1"),
            direct = FALSE, indirect = FALSE,
            ref = "first", data = dat
            
check_transitivity(modm)

modnet <- msc(scores = c("a", "b", "c", "d",),
            cohort = "cohort", outcome = "mort3", subjid = "id",
            fn = list(`logit AUC` = c_statistic, 
            `calib slope` = calibration_slope), 
            model = "consistency", 
            direct = TRUE, indirect = TRUE,
            ref = "first", data = dat)
            
check_consistency(modnet)}
}
