% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{consistency}
\alias{consistency}
\alias{inconsistency}
\title{MSC network meta-analysis models}
\usage{
consistency(x, mods = NULL, ...)

inconsistency(x, mods = NULL, ...)
}
\arguments{
\item{x}{A set of aggregated performance score data, as computed with the \code{\link{aggregate_performance}} function.}

\item{mods}{Which moderators should be included in the model? These can be any of the same moderators included in \code{\link{get_bs_samples}}, all others will be ignored.}

\item{...}{Any other arguments will be passed directly to \code{\link[metafor]{rma.mv}}}
}
\value{
The results of an \code{\link[metafor]{rma.mv}} multivariate/multilevel Linear (mixed-effects) model (class "rma.mv"), with the following additional components, which can be passed along to other functions:
#' \describe{
  \item{reference}{name of reference score}
  \item{scores}{vector of scores analysed}
  \item{model}{If \code{\link{consistency}} was used, "consistency", else "inconsistency"}
  \item{performance}{Label of performance, as used in previous functions}
}
}
\description{
MSC network meta-analysis models
}
\details{
The consistency and inconsistency models are those found in \href{https://doi.org/10.1186/s12874-016-0184-5}{Law et al 2016}: 

Specifically, we fit one of two models using "Jackson's Model", as described in the paper, which differ only in their random effects:
\describe{
  \item{consistency}{First item} random contrast within study
  \item{inconsistency}{Second item} random contrast within study, and random contrast within design
}
}
\section{Functions}{
\itemize{
\item \code{inconsistency}: Estimate differences in score performance using inconsistency model
}}

\examples{
dat <- msc_sample_data()
bssamp <- get_bs_samples(dat, id, study, outcome, n.samples = 10, 
                  scores = c("a", "b", "c", "d", "e", "f"), 
                  moderators = c("age", "female", "x1"))
perf <- compute_performance(bssamp, fn = calibration_slope, lbl = "CS")
agg <- aggregate_performance(perf)
consistency(agg)
agg.c <- aggregate_performance(perf, "c")
consistency(agg.c)
}
\seealso{
Law, M.; Jackson, D.; Turner, R.; Rhodes, K. & Viechtbauer, W. Two new methods to fit models for network meta-analysis with random inconsistency effects BMC Medical Research Methodology, 2016, 16, 87.
}
