% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrmat.R
\name{contrmat}
\alias{contrmat}
\title{Construct contrast matrix}
\usage{
contrmat(trt1, trt2, ref, sc = NULL)
}
\arguments{
\item{trt1}{vector of the 1st treatment / score in the contrast}

\item{trt2}{vector of the 2nd treatment / score in the contrast}

\item{ref}{The name of the reference treatment / score (character)}

\item{sc}{A vector containing the full list of scores considered, so that the scores are not put into alphabetical order. This keeps the order of the scores in later network meta-analysis models the same as in other places.}
}
\value{
A design matrix
}
\description{
Constructs a contrast matrix.
}
\examples{
contrmat(letters[c(1, 1, 2, 2, 3)], letters[c(2, 3, 4, 5, 6)], "a", sc = letters[1:6])
contrmat(letters[c(1, 1, 2, 2, 3)], letters[c(2, 3, 4, 5, 6)], "a", sc = letters[6:1])
contrmat(letters[c(1, 1, 2, 2, 3)], letters[c(2, 3, 4, 5, 6)], "c", sc = letters[6:1])
}
\references{
This function has been adapted slightly from the supplementary material of Law et al 2016:
Law, M.; Jackson, D.; Turner, R.; Rhodes, K. & Viechtbauer, W. Two new methods to fit models for
network meta-analysis with random inconsistency effects BMC Medical Research Methodology, 2016, 16, 87. \href{doi://10.1186/s12874-016-0184-5}
}
