% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff.R
\name{get_diff}
\alias{get_diff}
\title{Calculate differences between scores}
\usage{
get_diff(d)
}
\arguments{
\item{d}{A structured dataset, as calculated with \code{\link{aggregate_performance}}}
}
\value{
A new dataset with differences calculated
}
\description{
Calculate differences between scores
}
