% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k.R
\name{get_k}
\alias{get_k}
\title{Find number of non-missing scores}
\usage{
get_k(x)
}
\arguments{
\item{x}{A string of scores, pasted together with \code{:}.}
}
\value{
Number of non-missing scores
}
\description{
Calculates the number of non-missing scores row-wise. Not intended to be
called by the user.
}
