% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance_statistics.R
\name{get_performance_statistics}
\alias{get_performance_statistics}
\alias{basic_summ}
\alias{summary.perfsumm}
\title{Get performance statistics}
\usage{
get_performance_statistics(
  scores,
  cohort,
  outcome,
  data,
  fn = list(AUC = c_statistic, `O/E` = oe_ratio)
)

basic_summ(x, dig = 2)

\method{summary}{perfsumm}(x, fun = basic_summ, ...)
}
\arguments{
\item{scores}{A vector of scores (i.e. variables in the dataset corresponding to calculated scores). The first score in the vector will be used as a comparator for all other scores.}

\item{cohort}{The name of the variable corresponding to cohort or study in the dataset.}

\item{outcome}{The name of the variable corresponding to the outcome.}

\item{data}{Name of dataset containing all other variables, with individual patient data, that is, 1 row per subject.}

\item{fn}{A named list of performance metrics.}
}
\description{
Get performance statistics
}
\section{Functions}{
\itemize{
\item \code{basic_summ()}: Compute summary statistics

\item \code{summary(perfsumm)}: Summarize performance statistics

}}
