% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get_ref}
\alias{get_ref}
\title{Find reference score}
\usage{
get_ref(x)
}
\arguments{
\item{x}{A set of scores, pasted together.}
}
\value{
The number of the reference score
}
\description{
Finds the first non-missing scores, and selects it as the "reference" score,
for use in defining design matrices. Note intended to be called by the user.
}
\examples{
x <- paste(c(0.025, 0.05, NA, 0.1), sep = ":")
y <- paste(c(NA, 0.01, 0.02, 0.05), sep = ":")
get_ref(x)
get_ref(y)
}
