% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_table.R
\name{league_table}
\alias{league_table}
\title{Create league table}
\usage{
league_table(
  object,
  low = c("direct", "indirect", "network")[1],
  upp = c("direct", "indirect", "network")[2],
  dig = 2,
  mid = " to ",
  brackets = c("(", ")")
)
}
\arguments{
\item{object}{A model from \code{msc()}.}

\item{low}{Should the evidence in the lower triangle come from \code{"direct"} (default), \code{"indirect"}, or \code{"network"} evidence. It is sufficient to write the first letter.}

\item{dig}{Number of significant digits to format the results.}

\item{mid}{Confidence intervals are written as \code{brackets[1]} [lower bound] \code{mid} [upper bound] \code{brackets[2]}. (Default \code{" to "})}

\item{brackets}{Confidence intervals are written as \code{brackets[1]} [lower bound] \code{mid} [upper bound] \code{brackets[2]}. (Default \code{c("(", ")")})}

\item{upper}{Should the evidence in the upper triangle come from \code{"direct"}, \code{"indirect"} (default), or \code{"network"} evidence. It is sufficient to write the first letter.}
}
\value{
A matrix of performance comparisons, by pair of scores
}
\description{
Create league table
}
\examples{
dat <- msc_sample_data()
mod <- msc(scores = c("a", "b", "c", "d", "e"),
         cohort = "study", outcome = "outcome", subjid = "id", 
         data = dat, direct = TRUE, indirect = TRUE,
         ref = "all", model = "inconsistency")


}
