% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc_sample_data.R
\name{msc_sample_data}
\alias{msc_sample_data}
\title{Make MSC sample data set}
\usage{
msc_sample_data(n.cohorts = 30)
}
\arguments{
\item{n.cohorts}{Number of cohorts in the data set (default 30)}
}
\value{
A sample data set (tibble)
}
\description{
Creates a sample data set to test out the \code{\link{mscpredmodel}} package.
As the dataset is randomly generated, and it may be desired to make other
sample datasets (for example, with more missing observations, or different
}
\examples{
dat <- msc_sample_data(n.cohorts = 30)
}
