% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contrmat}
\alias{contrmat}
\alias{compute_performance}
\alias{bootfn}
\alias{get_contrasts}
\alias{get_perf_diff}
\alias{get_design}
\alias{get_mm}
\title{Various internal utility functions}
\usage{
contrmat(trt1, trt2, ref, sc = NULL)

compute_performance(
  x,
  k = k,
  sc = scores,
  fn = perffn,
  observed = outcome,
  missing_threshold = 0.8
)

bootfn(
  x,
  R = n.boot,
  newseed = NULL,
  k = k,
  sc = scores,
  pfn = perf_fn,
  missing_threshold = 0.8,
  observed = outcome
)

get_contrasts(x)

get_perf_diff(x)

get_design(x)

get_mm(x)
}
\arguments{
\item{trt1}{vector of the 1st treatment / score in the contrast}

\item{trt2}{vector of the 2nd treatment / score in the contrast}

\item{ref}{The name of the reference treatment / score (character)}

\item{sc}{A vector containing the full list of scores considered, so that the scores are not put into alphabetical order. This keeps the order of the scores in later network meta-analysis models the same as in other places.}

\item{x}{A variety of possible inputs, depends on the function}

\item{s1}{name of the 1st score in the contrast}

\item{s2}{name of the 2nd score in the contrast}

\item{s}{vector of scores to be kept}

\item{dl}{design.levels}

\item{to.check}{Name of score to be checked}
}
\value{
A design matrix

A new dataset with performance measures

A variance matrix

A vector

A vector

A matrix

A matrix
}
\description{
Various internal utility functions
}
\details{
This function has been adapted slightly from the supplementary material of  \href{https://doi.org/10.1186/s12874-016-0184-5}{Law et al 2016}:
}
\section{Functions}{
\itemize{
\item \code{contrmat()}: Make design matrix

\item \code{compute_performance()}: Calculate differences between performance measures

\item \code{bootfn()}: Calculate variance matrix of differences in performance measures, after bootstrap sampling

\item \code{get_contrasts()}: Get vector of which scores are compared

\item \code{get_perf_diff()}: Calculate difference in performance measure compared to first score

\item \code{get_design()}: Get design of study

\item \code{get_mm()}: Get model matrix

}}
\seealso{
Law, M.; Jackson, D.; Turner, R.; Rhodes, K. & Viechtbauer, W. Two new methods to fit models for network meta-analysis with random inconsistency effects BMC Medical Research Methodology, 2016, 16, 87.
}
