% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_performance.R
\name{aggregate_performance}
\alias{aggregate_performance}
\alias{print.mscagg}
\title{Aggregate performance measures by cohort}
\usage{
aggregate_performance(perf.estimates, reference = NULL,
  design.levels = LETTERS, fn.mods = NULL)

\method{print}{mscagg}(x, ...)
}
\arguments{
\item{perf.estimates}{A set of performance estimates of class \code{mscraw}, as computed by \code{\link{compute_performance}}}

\item{reference}{The name of the reference score (default NULL, the first score is the reference). This will be the reference level of the scores in any later models.}

\item{design.levels}{A character vector of alternate short names for the scores, to be used in naming the designs. Default is \code{\link{LETTERS}}, so that a possible design would be \code{AB} or \code{ACD}, instead of \code{score1score2} or \code{score1score3score4}. The design variable is used in the definition of the random effects. See \code{\link{consistency}}.}

\item{fn.mods}{A function used to summarize the moderators within cohorts. If NULL, the default mean(x, na.rm = TRUE) is used.}
}
\value{
A list with the following elements, of class \code{mscagg}
\describe{
  \item{cohorts}{List of cohorts}
  \item{yi}{Differences in score performance}
  \item{vi}{Variance-covariance matrix for yi}
  \item{contr}{Score contrasts}
  \item{design}{"Design" of the study (that is, which scores could be calculated)}
  \item{design.matrix}{Design matrix, corresponding to \code{contr}}
  \item{lbl}{A label for the analysis, which shows up in plots and output datasets}
  \item{fn}{The function used to compute performance, for example \code{\link{calibration_slope}}}
  \item{scores}{A vector of the scores used}
  \item{ref}{The reference score}
  \item{mods}{A vector of names of the moderators}
  \item{moderators}{A dataframe containing aggregated moderators}
}
}
\description{
Aggregate performance measures by cohort
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print basic aggregated performance measures
}}

\examples{
dat <- msc_sample_data()
bssamp <- get_bs_samples(dat, id, study, outcome, n.samples = 10, scores = c("a", "b", "c", "d", "e", "f"), moderators = c("age", "female", "x1"))
perf <- compute_performance(bssamp, fn = calibration_slope, lbl = "CS")
agg <- aggregate_performance(perf)
agg
}
