% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_slope.R
\name{calibration_slope}
\alias{calibration_slope}
\title{Compute calibration slope}
\usage{
calibration_slope(bss, fm)
}
\arguments{
\item{bss}{A set of bootstrap samples, stratified by cohort, as computed by \code{\link{get_bs_samples}}. The full bootstrap data is called within the function as \code{analysis(bss)}. See \code{\link{?rsample::bootstraps}} for more details.}

\item{fm}{The formula that will be called by the model, of the form \code{outcome ~ score} (character).}
}
\value{
A single calibration slope (numeric)
}
\description{
Compute calibration slope
}
