% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_transitivity}
\alias{check_transitivity}
\alias{check_homogeneity}
\alias{check_consistency}
\title{Check assumptions of transitivity, consistency and homogeneity}
\usage{
check_transitivity(ag, graph = FALSE)

check_homogeneity(mod, dig = 3)

check_consistency(ps, mtype = c("consistency", "inconsistency")[1])
}
\arguments{
\item{ag}{A set of aggregated performance scores}

\item{graph}{Should a graph of outcomes versus moderators be printed?}

\item{mod}{A \code{\link{consistency}} or \code{inconsistency} model.}

\item{dig}{Number of digits to be printed, see \code{\link{format.pval}}.}

\item{ps}{A set of raw performance estimates, from \code{\link{compute_performance}}}

\item{mtype}{Type of model (default "consistency", else "inconsistency"). It is sufficient to write \code{"c"} or \code{"i"}.}
}
\value{
A tibble containing results of linear regression models of the effect size (outcome) against the cohort-specific average moderator value.

A vector containing tau-square, the Q likelihood ratio statistic, its degrees of freedom and its p-value.

A graph of the direct and indirect estimates for the various differences in score performance.
}
\description{
Check assumptions of transitivity, consistency and homogeneity
}
\section{Functions}{
\itemize{
\item \code{check_transitivity}: Check assumption of transitivity

\item \code{check_homogeneity}: Check assumption of homogeneity

\item \code{check_consistency}: Check assumption of consistency
}}

\examples{
dat <- msc_sample_data()
bssamp <- get_bs_samples(dat, id, study, outcome, n.samples = 8, 
                  scores = c("a", "b", "c", "d", "e", "f"), 
                  moderators = c("age", "female", "x1"))
perf <- compute_performance(bssamp, fn = calibration_slope, lbl = "CS")
agg <- aggregate_performance(perf)
check_transitivity(agg, graph = FALSE)
check_homogeneity(consistency(agg))
\dontrun{check_consistency(perf)}
}
