% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_performance.R
\name{compute_performance}
\alias{compute_performance}
\alias{print.mscraw}
\alias{summary.mscraw}
\title{Computes performance measures for prognostic scores on a set of bootstrap samples}
\usage{
compute_performance(bs.sample, fn = calibration_slope, lbl = NULL)

\method{print}{mscraw}(x, ...)

\method{summary}{mscraw}(x, nonpar = TRUE, NArm = TRUE, ...)
}
\arguments{
\item{bs.sample}{A set of cohort-stratified bootstrap samples, as generated with \code{\link{get_bs_samples}}}

\item{fn}{The name of a function to compute performance measures. For example, \code{\link{calibration_slope}}. See \strong{details}.}

\item{lbl}{A label to describe the performance measure computed. For example \code{"calibration slope"}. This label is used in subsequent plot functions, and appears in tables of regression coefficients of. Useful if you want to compute several performance measures, or the same performance measure on different subsets of data.}
}
\value{
A list with 3 elements, of class \code{msc_raw}:
\describe{
  \item{working.estimates}{A tibble containing a set of "working estimates", the raw performance measures, before they have been aggregated or contrasts have been computed.}
  \item{scores}{Names of the scores as given in \code{\link{get_bs_samples}}}
  \item{formulas}{vector of formulas)}
  \item{fn}{The function definition used to compute performance}
  \item{lbl}{The label given in the arguments.}
}
The results of \code{\link{compute_performance}} have suitable \code{print} and \code{summary} methods.
}
\description{
Computes performance measures for prognostic scores on a set of bootstrap samples
}
\details{
The function to compute performance measures, \code{fn} requires two arguments:
\describe{ 
\item{bss}{The name of the bootstrap sample. The full bootstrap data is called within the function as \code{analysis(bss)}. See \code{\link{?rsample::bootstraps}} for more details.}
  \item{fn}{The formula that will be called by the model, of the form \code{outcome ~ score} (character).}
}
and outputs a single numeric value. Using \code{\link{possibly}}, \code{\link{compute_performance}} assigns a value of \code{NA} if there is an error.
}
\examples{
dat <- msc_sample_data()
bssamp <- get_bs_samples(dat, id, cohort, outcome, n.samples = 10, a, b, d, e, f)
perf <- compute_performance(bssamp, fn = calibration_slope, lbl = "CS")
print(perf)
summary(perf)
}
