% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bs_samples.R
\name{get_bs_samples}
\alias{get_bs_samples}
\title{Generates bootstrap samples}
\usage{
get_bs_samples(data, id, cohorts, outcome, n.samples = 1000, ...)
}
\arguments{
\item{data}{Dataset containing subject identifiers, names of cohorts, outcome, and each of the scores}

\item{id}{name of variable with subject identifiers}

\item{cohorts}{name of variable containing cohorts (the stratification factor, can be a factor, character or numeric)}

\item{outcome}{name of outcome variable (binomial)}

\item{n.samples}{Number of bootstrap samples to be generated within each cohort (default 1000)}

\item{...}{Any other arguments are taken to be the names of the variables containinng the scores. The order the scores are here is used in all other functions.}
}
\value{
A list with 3 elements:
\describe{
  \item{bs.sample}{A tibble containing the stratified bootstrap samples}
  \item{scores}{Names of the scores as given in \code{...} in the function call}
  \item{formulas}{vector of formulas that can be used in any function which calculates performance measures (e.g. \code{\link{calibration_slope}})}
}
}
\description{
Generates bootstrap samples
}
\examples{
dat <- msc_sample_data()
get_bs_samples(dat, id, cohort, outcome, n.samples = 10, a, b, d, e, f)
}
