% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bs_samples.R
\name{get_bs_samples}
\alias{get_bs_samples}
\title{Generates bootstrap samples}
\usage{
get_bs_samples(data, id, cohort, outcome, n.samples = 1000,
  scores = NULL, moderators = NULL)
}
\arguments{
\item{data}{Dataset containing subject identifiers, names of cohorts, outcome, and each of the scores}

\item{id}{name of variable with subject identifiers}

\item{cohort}{name of variable containing cohorts (the stratification factor, can be a factor, character or numeric)}

\item{outcome}{name of outcome variable (binomial)}

\item{n.samples}{Number of bootstrap samples to be generated within each cohort (default 1000)}

\item{scores}{A vector of variable names that are scores (or their predicted probabilities). The order the scores are here is used in all other functions.}

\item{moderators}{A vector of variable names that are moderators, that is, covariates which could affect the differences in score performance. We save these variables here, to be aggregated later in \code{\link{aggregate_performance}}. They can then be entered as covariates in \code{\link{msc_direct}}, \code{\link{consistency}}, or any other models.}
}
\value{
A list with 3 elements:
\describe{
  \item{bs.sample}{A data.frame containing the stratified bootstrap samples, split by cohort}
  \item{orig.sample}{A data.frame containing the orignal data, split by cohort}
  \item{scores}{Unique names of the scores which are also in the dataset}
  \item{mods}{A vector of moderators which may be included in subsequent models}
  \item{formulas}{vector of formulas that can be used in any function which calculates performance measures (e.g. \code{\link{calibration_slope}})}
}
}
\description{
Generates bootstrap samples
}
\examples{
dat <- msc_sample_data()
set.seed(12345)
get_bs_samples(dat, id, study, outcome, scores = letters[1:4], moderators = "age")
get_bs_samples(dat, id, "study", outcome, scores = letters[1:4], moderators = "age")
}
