% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc.R
\name{msc_full}
\alias{msc_full}
\alias{msc_network}
\alias{msc_indirect}
\alias{msc_direct}
\alias{print.msc}
\alias{plot.msc}
\title{Full, direct and indirect network results for MSC}
\usage{
msc_full(ps, mtype = c("consistency", "inconsistency")[1],
  verbose = FALSE, ...)

msc_network(ps, mtype = c("consistency", "inconsistency")[1], ...)

msc_indirect(ps, mtype = c("consistency", "inconsistency")[1],
  verbose = FALSE, ...)

msc_direct(ps, mtype = c("consistency", "inconsistency")[1],
  verbose = FALSE, ...)

\method{print}{msc}(x, ...)

\method{plot}{msc}(x, compare_to = NULL, newlabels = NULL)
}
\arguments{
\item{ps}{A set of raw performance estimates, see \code{\link{compute_performance}}. Since we have to compute a range of different aggregated performance measures, we start here with the raw performance estimates.}

\item{mtype}{Type of model (default "consistency", else "inconsistency"). It is sufficient to write \code{"c"} or \code{"i"}.}

\item{verbose}{If TRUE, results of each model will be printed (default FALSE)}

\item{...}{Other arguments to be passed to \code{\link[metafor]{rma.mv}}. See also \code{\link{consistency}}.}

\item{compare_to}{If specified, only comparisons to \code{compare_to} are plotted.}

\item{newlabels}{A new vector of labels (character) for the scores can be used instead of the current vector of score names.}

\item{...}{In the print function, other options may be specified. These are passed to \code{\link[tibble]{print.tbl}}.}
}
\value{
A list of class \code{msc}, with the following components:
\describe{
  \item{table}{A tibble containing the model results for each pair}
  \item{models}{A list with all the individual pairwise models results (\code{\link[metafor]{rma.mv}})}
  \item{type}{"Full", "indirect" or "direct"?}
  \item{performance}{Label of performance score, as given in, for example, \code{\link{compute_performance}}.}
}
}
\description{
Compute all pairwise comparisons for the full network, all direct comparisons, or all indirect.
}
\section{Functions}{
\itemize{
\item \code{msc_full}: Compute all pairwise comparisons in full network of evidence, as well as direct and indirect comparisons

\item \code{msc_network}: Compute all pairwise comparisons in the full network of evidence

\item \code{msc_indirect}: Compute all pairwise indirect comparisons

\item \code{msc_direct}: Compute all pairwise indirect comparisons
}}

\examples{
dat <- msc_sample_data()
bssamp <- get_bs_samples(dat, id, cohort, outcome, n.samples = 10, a, b, c, d, e)
perf <- compute_performance(bssamp, fn = calibration_slope, lbl = "CS")
msc_indirect(perf, mtype = "inconsistency")
msc_direct(perf, mtype = "inconsistency")
msc_network(perf, mtype = "inconsistency")
full <- msc_full(perf, mtype = "inconsistency")
plot(full)

}
