% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscpredmodel.R
\docType{package}
\name{mscpredmodel}
\alias{mscpredmodel}
\alias{mscpredmodel-package}
\title{mscpredmodel: Multiple Score Comparison of Prediction Models with NMA}
\description{
The mscpredmodel package provides a number of functions to facilitate
multiple score comparison (MSC), a network meta-analytic approach to 
comparing multiple prediction models or prognostic scores using 
individual patient data (IPD).
}
