% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscpredmodel.R
\docType{package}
\name{mscpredmodel}
\alias{mscpredmodel-package}
\alias{mscpredmodel}
\title{mscpredmodel: Multiple Score Comparison of Prediction Models}
\description{
The mscpredmodel package provides a number of functions to
facilitate multiple score comparison (MSC), a network meta-analytic 
approach to comparing multiple prediction models or prognostic scores using 
individual patient data (IPD).
}
\section{Main Functions}{

The main mscpredmodel function is:
\describe{
  \item{\code{\link{msc}}}{Compare performance of prediction models, can calculate any model for full network comparisons, or direct or indirect comparisons, and can adjust for moderators.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/srhaile/mscpredmodel}
  \item Report bugs at \url{http://github.com/srhaile/mscpredmodel/issues}
}

}
\author{
\strong{Maintainer}: Sarah Haile \email{sarah.haile@uzh.ch}

}
