% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contrmat}
\alias{contrmat}
\alias{get_diff}
\alias{get_ref}
\alias{get_scores}
\alias{get_design}
\alias{get_est}
\alias{get_var}
\alias{get_direct}
\alias{get_indirect}
\alias{check_combn}
\title{Various utility functions}
\usage{
contrmat(trt1, trt2, ref, sc = NULL)

get_diff(x)

get_ref(x)

get_scores(x)

get_design(x)

get_est(x, s)

get_var(x, s)

get_direct(x, s1, s2)

get_indirect(x, s1, s2)

check_combn(x, to.check = s1)
}
\arguments{
\item{trt1}{vector of the 1st treatment / score in the contrast}

\item{trt2}{vector of the 2nd treatment / score in the contrast}

\item{ref}{The name of the reference treatment / score (character)}

\item{sc}{A vector containing the full list of scores considered, so that the scores are not put into alphabetical order. This keeps the order of the scores in later network meta-analysis models the same as in other places.}

\item{x}{A variety of possible inputs, depends on the function}

\item{s1}{vector of the 1st score in the contrast}

\item{s2}{vector of the 2nd score in the contrast}

\item{to.check}{Name of score to be checked}
}
\value{
A design matrix

A new dataset with differences calculated

The number of the reference score

The names of the observed scores

A string of the design (combination of scores) of the cohort

a vector of performance estimates

a variance matrix

a data.frame

a data.frame

TRUE/FALSE
}
\description{
Various utility functions
}
\details{
This function has been adapted slightly from the supplementary material of  \href{https://doi.org/10.1186/s12874-016-0184-5}{Law et al 2016}: 

Law, M.; Jackson, D.; Turner, R.; Rhodes, K. & Viechtbauer, W. Two new methods to fit models for network meta-analysis with random inconsistency effects BMC Medical Research Methodology, 2016, 16, 87.
}
\section{Functions}{
\itemize{
\item \code{contrmat}: Make design matrix

\item \code{get_diff}: Calculate differences between performance measures

\item \code{get_ref}: Get reference score

\item \code{get_scores}: Get non-missing scores

\item \code{get_design}: Get design of the cohort

\item \code{get_est}: Get estimated performance

\item \code{get_var}: Get variance estimated performance

\item \code{get_direct}: Find direct (head-to-head) comparisons

\item \code{get_indirect}: Find indirect comparisons

\item \code{check_combn}: Check if score is still in any of the cohorts?
}}

